Book examples from 20th chapter.

+++ src/homework_18/Ani_Kovalenko/GenericMap.java
@@ -76,7 +76,7 @@ public class GenericMap<K extends User, V extends String> {
     }
 
     private void checkLoadFactor() {
-        double notNullBucketCount = 0.0;
+        int notNullBucketCount = 0;
         for (int i = 0; i < buckets.length; i++) {
             if (buckets[i] != null) {
                 notNullBucketCount++;
